Move linear maps out of knossos.h
